Update C# tree-sitter grammar to support C# 12 and 13
Upgrade vendored grammar from v0.23.1 to tree-sitter-c-sharp master
(commit 8836663) which includes PR #364 adding semicolon-terminated
type declarations and 46 additional fixes. This resolves 34 parse
failures in GrandNode2 caused by C# 12 bodyless types like
`class Foo : Bar;` and ensures controller base class inheritance
chains are correctly recorded.

Restructure csharp/ to use tree_sitter/ subdirectory for headers
(matching the PHP grammar pattern) and bump LANGUAGE_VERSION 14→15.
