+++ csharp/binding.go
@@ -1,13 +1,16 @@
 package csharp
 
 //#cgo CFLAGS: -Wno-trigraphs
-//#include "parser.h"
+//#include "tree_sitter/parser.h"
 //TSLanguage *tree_sitter_c_sharp();
 import "C"
 import (
 	"unsafe"
 
 	sitter "github.com/api-extraction-examples/go-tree-sitter"
+	// Blank import to ensure go mod vendor copies the tree_sitter directory
+	// which contains header files required by the C code.
+	_ "github.com/api-extraction-examples/go-tree-sitter/csharp/tree_sitter"
 )
 
 // GetLanguage returns a grammar for C# language.
