+++ csharp/binding_test.go
@@ -3,6 +3,7 @@ package csharp_test
 import (
 	"context"
 	"os"
+	"strings"
 	"testing"
 
 	sitter "github.com/api-extraction-examples/go-tree-sitter"
@@ -29,9 +30,115 @@ func TestGrammar2(t *testing.T) {
 
 	n, err := sitter.ParseCtx(context.Background(), content, csharp.GetLanguage())
 	assert.NoError(err)
-	expected := "(compilation_unit (using_directive (qualified_name qualifier: (qualified_name qualifier: (identifier) name: (identifier)) name: (identifier))) (using_directive (qualified_name qualifier: (qualified_name qualifier: (identifier) name: (identifier)) name: (identifier))) (using_directive (qualified_name qualifier: (identifier) name: (identifier))) (using_directive (qualified_name qualifier: (qualified_name qualifier: (identifier) name: (identifier)) name: (identifier))) (using_directive (qualified_name qualifier: (qualified_name qualifier: (identifier) name: (identifier)) name: (identifier))) (using_directive (qualified_name qualifier: (qualified_name qualifier: (identifier) name: (identifier)) name: (identifier))) (file_scoped_namespace_declaration name: (qualified_name qualifier: (qualified_name qualifier: (qualified_name qualifier: (identifier) name: (identifier)) name: (identifier)) name: (identifier))) (class_declaration (attribute_list (attribute name: (identifier))) (attribute_list (attribute name: (identifier) (attribute_argument_list (attribute_argument (string_literal (string_literal_content)))))) (attribute_list (attribute name: (identifier) (attribute_argument_list (attribute_argument (string_literal (string_literal_content)))))) (attribute_list (attribute name: (identifier) (attribute_argument_list (attribute_argument (string_literal (string_literal_content)))))) (preproc_if_in_attribute_list condition: (unary_expression argument: (identifier)) (attribute_list (attribute name: (identifier) (attribute_argument_list (attribute_argument (assignment_expression left: (identifier) right: (boolean_literal))))))) (modifier) name: (identifier) (base_list (identifier)) body: (declaration_list (field_declaration (modifier) (modifier) (variable_declaration type: (identifier) (variable_declarator name: (identifier)))) (constructor_declaration (modifier) name: (identifier) parameters: (parameter_list (parameter type: (identifier) name: (identifier))) body: (block (expression_statement (assignment_expression left: (identifier) right: (identifier))))) (comment) (comment) (comment) (method_declaration (attribute_list (attribute name: (identifier))) (attribute_list (attribute name: (identifier) (attribute_argument_list (attribute_argument (string_literal (string_literal_content)))))) (attribute_list (attribute name: (identifier) (attribute_argument_list (attribute_argument (string_literal (string_literal_content)))))) (attribute_list (attribute name: (identifier) (attribute_argument_list (attribute_argument (string_literal (string_literal_content)))))) (modifier) returns: (identifier) name: (identifier) parameters: (parameter_list (parameter type: (identifier) name: (identifier) (default_expression))) body: (block (return_statement (invocation_expression function: (identifier) arguments: (argument_list (argument (identifier))))))) (comment) (comment) (comment) (method_declaration (attribute_list (attribute name: (identifier))) (attribute_list (attribute name: (identifier) (attribute_argument_list (attribute_argument (string_literal (string_literal_content)))))) (attribute_list (attribute name: (identifier) (attribute_argument_list (attribute_argument (string_literal (string_literal_content)))))) (attribute_list (attribute name: (identifier) (attribute_argument_list (attribute_argument (string_literal (string_literal_content)))))) (modifier) (modifier) returns: (identifier) name: (identifier) parameters: (parameter_list (parameter type: (identifier) name: (identifier) (default_expression))) body: (block (if_statement condition: (member_access_expression expression: (member_access_expression expression: (identifier) name: (identifier)) name: (identifier)) consequence: (block (expression_statement (await_expression (invocation_expression function: (member_access_expression expression: (identifier) name: (identifier)) arguments: (argument_list (argument (identifier)) (argument (identifier)) (argument (identifier))))))) alternative: (block (expression_statement (assignment_expression left: (member_access_expression expression: (member_access_expression expression: (identifier) name: (identifier)) name: (identifier)) right: (member_access_expression expression: (identifier) name: (identifier)))))))))))"
+	expected := "(compilation_unit (using_directive (qualified_name qualifier: (qualified_name qualifier: (identifier) name: (identifier)) name: (identifier))) (using_directive (qualified_name qualifier: (qualified_name qualifier: (identifier) name: (identifier)) name: (identifier))) (using_directive (qualified_name qualifier: (identifier) name: (identifier))) (using_directive (qualified_name qualifier: (qualified_name qualifier: (identifier) name: (identifier)) name: (identifier))) (using_directive (qualified_name qualifier: (qualified_name qualifier: (identifier) name: (identifier)) name: (identifier))) (using_directive (qualified_name qualifier: (qualified_name qualifier: (identifier) name: (identifier)) name: (identifier))) (file_scoped_namespace_declaration name: (qualified_name qualifier: (qualified_name qualifier: (qualified_name qualifier: (identifier) name: (identifier)) name: (identifier)) name: (identifier))) (class_declaration (attribute_list (attribute name: (identifier))) (attribute_list (attribute name: (identifier) (attribute_argument_list (attribute_argument (string_literal (string_literal_content)))))) (attribute_list (attribute name: (identifier) (attribute_argument_list (attribute_argument (string_literal (string_literal_content)))))) (attribute_list (attribute name: (identifier) (attribute_argument_list (attribute_argument (string_literal (string_literal_content)))))) (preproc_if_in_attribute_list condition: (unary_expression argument: (identifier)) (attribute_list (attribute name: (identifier) (attribute_argument_list (attribute_argument name: (identifier) (boolean_literal)))))) (modifier) name: (identifier) (base_list (identifier)) body: (declaration_list (field_declaration (modifier) (modifier) (variable_declaration type: (identifier) (variable_declarator name: (identifier)))) (constructor_declaration (modifier) name: (identifier) parameters: (parameter_list (parameter type: (identifier) name: (identifier))) body: (block (expression_statement (assignment_expression left: (identifier) right: (identifier))))) (comment) (comment) (comment) (method_declaration (attribute_list (attribute name: (identifier))) (attribute_list (attribute name: (identifier) (attribute_argument_list (attribute_argument (string_literal (string_literal_content)))))) (attribute_list (attribute name: (identifier) (attribute_argument_list (attribute_argument (string_literal (string_literal_content)))))) (attribute_list (attribute name: (identifier) (attribute_argument_list (attribute_argument (string_literal (string_literal_content)))))) (modifier) returns: (identifier) name: (identifier) parameters: (parameter_list (parameter type: (identifier) name: (identifier) (default_expression))) body: (block (return_statement (invocation_expression function: (identifier) arguments: (argument_list (argument (identifier))))))) (comment) (comment) (comment) (method_declaration (attribute_list (attribute name: (identifier))) (attribute_list (attribute name: (identifier) (attribute_argument_list (attribute_argument (string_literal (string_literal_content)))))) (attribute_list (attribute name: (identifier) (attribute_argument_list (attribute_argument (string_literal (string_literal_content)))))) (attribute_list (attribute name: (identifier) (attribute_argument_list (attribute_argument (string_literal (string_literal_content)))))) (modifier) (modifier) returns: (identifier) name: (identifier) parameters: (parameter_list (parameter type: (identifier) name: (identifier) (default_expression))) body: (block (if_statement condition: (member_access_expression expression: (member_access_expression expression: (identifier) name: (identifier)) name: (identifier)) consequence: (block (expression_statement (await_expression (invocation_expression function: (member_access_expression expression: (identifier) name: (identifier)) arguments: (argument_list (argument (identifier)) (argument (identifier)) (argument (identifier))))))) alternative: (block (expression_statement (assignment_expression left: (member_access_expression expression: (member_access_expression expression: (identifier) name: (identifier)) name: (identifier)) right: (member_access_expression expression: (identifier) name: (identifier)))))))))))"
 	assert.Equal(
 		expected,
 		n.String(),
 	)
+}
+
+// TestCSharp12_SemicolonTerminatedTypes verifies that C# 12 semicolon-terminated
+// (bodyless) type declarations parse without errors. This was the primary fix
+// from upstream PR #364 (commit d13ccdd).
+func TestCSharp12_SemicolonTerminatedTypes(t *testing.T) {
+	assert := assert.New(t)
+
+	code := `public class BasePluginController : ControllerBase;`
+	n, err := sitter.ParseCtx(context.Background(), []byte(code), csharp.GetLanguage())
+	assert.NoError(err)
+
+	ast := n.String()
+	assert.NotContains(ast, "ERROR", "semicolon-terminated class should parse without errors")
+	assert.Contains(ast, "class_declaration")
+	assert.Contains(ast, "base_list")
+}
+
+// TestCSharp12_SemicolonTerminatedVariants verifies semicolon-terminated syntax
+// for all type kinds: class, struct, record, record struct, enum, and interface.
+func TestCSharp12_SemicolonTerminatedVariants(t *testing.T) {
+	tests := []struct {
+		name     string
+		code     string
+		nodeType string
+	}{
+		{"class", "public class Foo : Bar;", "class_declaration"},
+		{"struct", "public struct Foo;", "struct_declaration"},
+		{"record", "public record Foo;", "record_declaration"},
+		{"record struct", "public record struct Foo;", "record_declaration"},
+		{"interface", "public interface IFoo;", "interface_declaration"},
+		{"enum", "public enum Foo;", "enum_declaration"},
+		{"generic class", "public class Foo<T> : Bar<T>;", "class_declaration"},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			n, err := sitter.ParseCtx(context.Background(), []byte(tt.code), csharp.GetLanguage())
+			assert.NoError(t, err)
+			ast := n.String()
+			assert.NotContains(t, ast, "ERROR", "should parse without errors: %s", tt.code)
+			assert.Contains(t, ast, tt.nodeType)
+		})
+	}
+}
+
+// TestCSharp12_PrimaryConstructors verifies primary constructor syntax on classes.
+func TestCSharp12_PrimaryConstructors(t *testing.T) {
+	assert := assert.New(t)
+
+	code := `public class UserService(ILogger logger, IRepository repo)
+{
+    public void Log(string message) => logger.Log(message);
+}`
+	n, err := sitter.ParseCtx(context.Background(), []byte(code), csharp.GetLanguage())
+	assert.NoError(err)
+
+	ast := n.String()
+	assert.NotContains(ast, "ERROR", "primary constructor should parse without errors")
+	assert.Contains(ast, "class_declaration")
+}
+
+// TestCSharp12_PrimaryConstructorWithBaseAndSemicolon verifies that a class with
+// a primary constructor, base invocation, and semicolon body parses correctly.
+func TestCSharp12_PrimaryConstructorWithBaseAndSemicolon(t *testing.T) {
+	assert := assert.New(t)
+
+	code := `public class DerivedService(ILogger logger) : BaseService(logger);`
+	n, err := sitter.ParseCtx(context.Background(), []byte(code), csharp.GetLanguage())
+	assert.NoError(err)
+
+	ast := n.String()
+	assert.NotContains(ast, "ERROR", "primary constructor with base and semicolon should parse without errors")
+	assert.Contains(ast, "class_declaration")
+	assert.Contains(ast, "base_list")
+}
+
+// TestCSharp12_FullFile parses the C# 12 test fixture and verifies no errors.
+func TestCSharp12_FullFile(t *testing.T) {
+	assert := assert.New(t)
+
+	content, err := os.ReadFile("testing/csharp12.cs")
+	assert.NoError(err)
+
+	n, err := sitter.ParseCtx(context.Background(), content, csharp.GetLanguage())
+	assert.NoError(err)
+
+	ast := n.String()
+	assert.NotContains(ast, "ERROR", "C# 12 test fixture should parse without errors")
+}
+
+// TestCSharp13_FullFile parses the C# 13 test fixture and verifies no errors.
+func TestCSharp13_FullFile(t *testing.T) {
+	assert := assert.New(t)
+
+	content, err := os.ReadFile("testing/csharp13.cs")
+	assert.NoError(err)
+
+	n, err := sitter.ParseCtx(context.Background(), content, csharp.GetLanguage())
+	assert.NoError(err)
+
+	ast := n.String()
+	// Count ERROR nodes if any — some C# 13 features may not be fully supported
+	errorCount := strings.Count(ast, "ERROR")
+	t.Logf("C# 13 parse errors: %d", errorCount)
+	// Partial properties may not be fully supported; allow a small number of errors
+	assert.LessOrEqual(errorCount, 2, "C# 13 test fixture should have at most 2 parse errors")
 }
