+++ csharp/scanner.c
@@ -1,6 +1,6 @@
-#include "alloc.h"
-#include "array.h"
-#include "parser.h"
+#include "tree_sitter/alloc.h"
+#include "tree_sitter/array.h"
+#include "tree_sitter/parser.h"
 
 #include <wctype.h>
