+++ csharp/tree_sitter/doc.go
@@ -1,0 +1,4 @@
+// Package tree_sitter contains C header files required by the C# grammar.
+// This file exists solely to ensure go mod vendor copies this directory.
+// See: https://github.com/golang/go/issues/26366
+package tree_sitter
